fresh approach to standardizer
A fresh start on the standardizer, uses openpyxl library to read raw excel sheet. Currently the script only deals with reading in the meta data and tries to output it to a csv file. Currently there is one error in the code: 

UnicodeEncodeError                        Traceback (most recent call last)
<ipython-input-31-9355f74c4cc5> in <module>()
     61                     'Emri i menaxherit t‘ prokurimit': ws['I18'].value,
     62                     'E-mail adresa': ws['I19'].value,
---> 63                     'Adresa e webit t‘ AK' : ws['I20'].value
     64                     })
     65 
//anaconda/lib/python2.7/csv.pyc in writerow(self, rowdict)
    150 
    151     def writerow(self, rowdict):
--> 152         return self.writer.writerow(self._dict_to_list(rowdict))
    153 
    154     def writerows(self, rowdicts):
UnicodeEncodeError: 'ascii' codec can't encode character u'\xeb' in position 17: ordinal not in range(128)

I'm not sure where I can add in a specification to encode using utf-8